moves addToLocal to career.logic relates #102
